html/canvas/element/wide-gamut-canvas/2d.color.space.p3.color.space.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.color.space.p3.color.space</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.color.space.p3.color.space</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

var color_style = 'rgb(50, 100, 150)';
var color_p3_f32 = [0.24304, 0.38818, 0.57227, 1.0];
var epsilon =  2;
ctx.fillStyle = color_style;
ctx.fillRect(0, 0, 10, 10);

var pixel = ctx.getImageData(5, 5, 1, 1, {colorSpace: "display-p3", storageFormat: "uint8"}).data;
var pixel_expected  = color_p3_f32;
pixel_expected = [pixel_expected[0] * 255,
                  pixel_expected[1] * 255,
                  pixel_expected[2] * 255,
                  pixel_expected[3] * 255];
var passed = true;
_assertSame(pixel.length, pixel_expected.length, "pixel.length", "pixel_expected.length");
for (var i = 0; i < pixel.length; i++)
{
  if (Math.abs(pixel[i] - pixel_expected[i]) > epsilon) {
    passed = false;
  }
}
_assertSame(passed, true, "passed", "true");


}, {colorSpace: "display-p3", pixelFormat: "uint8"});
</script>
